+++ changelog.md
@@ -1,3 +1,5 @@
 ### Changes and new features
 
 - Implemented option `ALL` for auxiliary blocks
+- Implemented singleplayer-only feature to get maps of the mapart without
+  building them in the world